Update layout to include stylesheet packs
The defaul setting uses stylesheet_link_tags which is used for the asset
pipeline. Since we're using webpacker, we want to use stylesheet_pack_tags so we
correctly include the relevant packs.

It might be that we want to use both, if we start adding more custom CSS. While
I'm not sure, I don't think that will be the case...